move player fixtures
