+++ tests/integration/conftest.py
@@ -22,9 +22,9 @@ from shvatka.clients.file_storage import FileStorage
 from shvatka.scheduler import Scheduler
 from shvatka.utils.key_checker_lock import KeyCheckerFactory
 from tests.fixtures.conftest import fixtures_resource_path  # noqa: F401
+from tests.fixtures.player import harry, hermione, ron, author  # noqa: F401
 from tests.fixtures.scn_fixtures import simple_scn  # noqa: F401
 from tests.mocks.config import DBConfig
-from tests.utils.player import harry, hermione, ron, author  # noqa: F401
 from tgbot.config.models.main import TgBotConfig
 from tgbot.main_factory import (
     create_dispatcher, create_scheduler, create_redis,
